requirements.txt
beautifulsoup4
bs4
mysql-connector
requests
scikit-learn
sklearn

